add alert, invoice_no and totalAmount in make_order and invoice_card.blade.php

+++ resources/views/order/store_recipe_data.blade.php
@@ -3,19 +3,22 @@
         const recipeCards = document.querySelectorAll('.recipe-data');
 
         recipeCards.forEach(card => {
-            card.addEventListener('click', function () {
+            card.addEventListener('click', function (event) {
+                event.preventDefault();
                 const recipeId = this.dataset.id;
                 const recipeName = this.dataset.name;
                 const recipeAmount = this.dataset.amount;
                 const recipeDiscount = this.dataset.discount;
                 const recipeImage = this.dataset.image;
 
+
                 const recipeData = {
                     id: recipeId,
                     name: recipeName,
                     amount: recipeAmount,
                     discount: recipeDiscount,
-                    image: recipeImage
+                    image: recipeImage,
+
                 };
 
                 recipeData.quantity = 1;
@@ -23,11 +26,28 @@
                 // Retrieve existing recipes from localStorage
                 let selectedRecipes = JSON.parse(localStorage.getItem('selectedRecipes')) || [];
 
-                // Add the new recipe to the array
-                selectedRecipes.push(recipeData);
+                // Check if the recipe ID already exists in localStorage
+                const existsInLocalStorage = selectedRecipes.some(r => r.id === recipeId);
 
-                localStorage.setItem('selectedRecipes', JSON.stringify(selectedRecipes));
-                window.location.href = this.dataset.url;
+                if (existsInLocalStorage) {
+                    Swal.fire({
+                        position: "center",
+                        icon: "warning",
+                        title: "This recipe is already in your order.",
+                        showConfirmButton: true,
+                        timer: 3000
+                    });
+                    return;
+                } else {
+                    // Add the new recipe to the array
+                    selectedRecipes.push(recipeData);
+
+                    // Update localStorage with the updated recipes array
+                    localStorage.setItem('selectedRecipes', JSON.stringify(selectedRecipes));
+
+                    // Redirect to the URL specified in dataset
+                    window.location.href = this.dataset.url;
+                }
             });
         });
     });
